jobs api refactored

+++ src/controllers/job.js
@@ -12,7 +12,7 @@ const getJobById = asyncHandler(async (req, res) => {
     throw new Error('Job not Found');
   }
 
-  res.json(job);
+  res.status(200).json(job);
 });
 
 const getJobs = asyncHandler(async (req, res) => {
@@ -21,13 +21,16 @@ const getJobs = asyncHandler(async (req, res) => {
   const total = await Job.countDocuments(filterQueries);
 
   if (page > Math.ceil(total / limit) && total > 0) {
+    res.status(404);
     throw new Error('Page not Found');
   }
+
+  const skipCount = (page - 1) * limit;
 
   const jobs = await Job.find(filterQueries)
     .collation({ locale: 'en', strength: 2 })
     .sort(sort || '-createdAt')
-    .skip((page - 1) * limit)
+    .skip(skipCount)
     .limit(+limit);
 
   res.status(200).json({
@@ -49,6 +52,8 @@ const getJobs = asyncHandler(async (req, res) => {
 });
 
 const createJob = asyncHandler(async (req, res) => {
+  req.body.createdBy = req.user._id;
+
   const job = await Job.create(req.body);
 
   res.status(201).json(job);
@@ -56,33 +61,48 @@ const createJob = asyncHandler(async (req, res) => {
 
 const updateJob = asyncHandler(async (req, res) => {
   const { id } = req.params;
+  const { _id: userId } = req.user;
+  req.body.updatedBy = userId;
 
-  const options = {
-    new: true,
-    runValidators: true,
-  };
-
-  const updatedJob = await Job.findByIdAndUpdate(id, req.body, options);
-
-  if (!updatedJob) {
-    res.status(404);
-    throw new Error('Job not Found');
-  }
-
-  res.status(200).json(updatedJob);
-});
-
-const deleteJob = asyncHandler(async (req, res) => {
-  const { id } = req.params;
-
-  const job = await Job.findByIdAndDelete(id);
+  const job = await Job.findById(id);
 
   if (!job) {
     res.status(404);
     throw new Error('Job not Found');
   }
 
-  res.status(204).json({ message: 'Job Deleted' });
+  if (userId.toString() !== job.createdBy.toString()) {
+    res.status(403);
+    throw new Error('Not Authorized to update this job');
+  }
+
+  const updatedJob = await Job.findByIdAndUpdate(id, req.body, {
+    new: true,
+    runValidators: true,
+  });
+
+  res.status(200).json(updatedJob);
+});
+
+const deleteJob = asyncHandler(async (req, res) => {
+  const { id } = req.params;
+  const { _id: userId } = req.user;
+
+  const job = await Job.findById(id);
+
+  if (!job) {
+    res.status(404);
+    throw new Error('Job not Found');
+  }
+
+  if (userId.toString() !== job.createdBy.toString()) {
+    res.status(403);
+    throw new Error('Not Authorized to delete this job');
+  }
+
+  await Job.findByIdAndDelete(id);
+
+  res.status(200).json({ message: 'Job Deleted' });
 });
 
 export { getJobById, getJobs, createJob, updateJob, deleteJob };
